+++ workflows/toast_so_sim.py
@@ -9,7 +9,7 @@ New users should either use `toast_so_sim`
 or `python -m sotodlib.workflows.toast_so_sim`.
 """
 
-from sotodlib.workflows.toast_so_sim import cli
+from sotodlib.toast.workflows.toast_so_sim import cli
 
 
 if __name__ == "__main__":
